[CHORE] MovieTable nowPlayingHeader 밖으로 빼냄
